moving makeBasicAuth into shared lib
